requirements.txt
dataclasses-json>=0.3.7
GitPython>=3.0.5
ipykernel
matplotlib>=3.1.2
numpy>=1.16.4
openpyxl>=3.0.3
pandas>=0.25.3
pipdeptree==0.13.2
psutil>=5.6.3
python-dateutil>=2.8.0
requests==2.22.0
scipy>=1.3.2
SQLAlchemy>=1.3.13
sty==1.0.0b11
